refactored protected route, created component instead of HOC
