+++ package.json
@@ -54,7 +54,7 @@
     "html-webpack-plugin": "5.3.2",
     "husky": "7.0.2",
     "jest": "27.3.1",
-    "jsdom": "InnerForm18.0.0",
+    "jsdom": "18.0.0",
     "jsdom-global": "3.0.2",
     "npm-run-all": "4.1.5",
     "path": "0.12.7",
@@ -82,6 +82,7 @@
     "react-dom": "17.0.2",
     "react-draft-wysiwyg": "1.14.7",
     "react-redux": "7.2.5",
+    "react-router": "5.2.1",
     "react-router-dom": "5.3.0",
     "redux-saga": "1.1.3",
     "styled-components": "5.3.1",
